requirements_api.txt
fastapi
flask
sse-starlette
